backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pandas
numpy
python-dateutil
pytesseract
Pillow
opencv-python
aiofiles
PyMuPDF
openpyxl
xlrd
gunicorn
httptools
reportlab
PyPDF2
easyocr
torch
torchvision
python-doctr[torch]
transformers
paddlepaddle
paddleocr
ultralytics
scikit-learn
scipy
matplotlib
Jinja2
lxml
xmltodict
cachetools
psutil
requests
pydantic
typing-extensions
httpx
starlette
anyio
sniffio
idna
certifi
motor
pymongo
python-dotenv
bcrypt
python-jose[cryptography]
passlib[bcrypt]
email-validator
